+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # caravaggio-react
 
-## 0.4.0
+## Next
+  - ⚠ Breaking changes
+  - Image can be null when using hook `useCaravaggioIfAvailable`
+  - "useCaravaggioImage" has been renamed to "useCarvaggio". the old name is deprecated
 
+## 0.4.0
   - Add explanation of parameters
   - `opt` is optional in Image component. If not passed, no transformation is applied and it behaves
     like a normal `img` tag.
@@ -10,5 +14,4 @@
 
 
 ## 0.3.9
-
   - Image url is correctly escaped